pyproject.toml
[tool.poetry]
name = "Rag app"
version = "0.1.0"
description = "rag application"
authors = ["dumball <archit1290@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "3.0.0"
Flask-Cors = "4.0.0"
Flask-SQLAlchemy = "3.1.1"
PyPDF2 = "3.0.1"
numpy = "1.26.4"
llama_cpp_python = "0.2.50"
faiss-cpu = "1.7.4"
#markers="python_implementation=='cpython' and not(platform_machine.startswith('arm'))"
#markers = "sys_platform == 'linux' and python_implementation == 'cpython' and not platform_machine.startswith('arm')"
#torch = { version = "2.0.1", markers = "python_implementation == 'cpython' and sys_platform == 'linux' and not platform_machine.startswith('arm')" }
#torch = { version = "2.0.1+cpu", markers = "sys_platform == 'linux' and platform_machine == 'x86_64'" }
torch = { version = "2.0.1", markers = "sys_platform == 'linux'" }
#torch = {version = "2.0.1", markers = "sys_platform == 'linux' and python_implementation == 'cpython' and not (platform_machine.startswith('arm'))"}
#torch = {version = "2.0.1", markers = "sys_platform == 'linux' and platform_python_implementation == 'CPython' and not platform_machine.startswith('arm')"}
#torch = {version = "2.0.1", markers = "sys_platform == 'linux' and platform_python_implementation == 'CPython' and not (platform_machine.startswith('arm'))"}
#torch = {version = "2.0.1", markers = "platform_machine == 'x86_64' and platform_python_implementation == 'CPython'"}
#torch = { version = "2.0.1", markers = "cpu" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

